Added inference script for old Model

+++ MobileNetV3_keras/src/generator.py
@@ -76,8 +76,10 @@ class DataGenerator(Sequence):
         return X, y
 
     def __data_generation(self, paths_batch_list, label_batch_list):
-        X = np.empty((self.batch_size, self.image_height, self.image_width, 3), dtype='float32')
-        y = np.empty((self.batch_size, len(label_batch_list[0])), dtype='float32')
+        X = np.empty((self.batch_size, self.image_height, self.image_width, 3))
+        X = X.astype('float32')
+        y = np.empty((self.batch_size, len(label_batch_list[0])))
+        y = y.astype('float32')
         #print (len(self.label_list[0]))
 
         for i, (path, label) in enumerate(zip(paths_batch_list, label_batch_list)):
